Updates driven by Long Line implementation
In addition to Long Line updates, this revision contains the seeds
of multi-tenant security.  This isn't working yet, but the comments
and tests show what has been tried and found wanting.

+++ TubsDataAccessLayer/Entities/Gen5Object.cs
@@ -64,11 +64,11 @@ namespace Spc.Ofp.Tubs.DAL.Entities
 
         public virtual FadType? AsLeft { get; set; }
 
-        public virtual int? Depth { get; set; }
+        public virtual decimal? Depth { get; set; }
 
-        public virtual int? Length { get; set; }
+        public virtual decimal? Length { get; set; }
 
-        public virtual int? Width { get; set; }
+        public virtual decimal? Width { get; set; }
 
         public virtual string BuoyNumber { get; set; }
